moved journal and photos over as components. Set  a ternary operater to handle which is shown based on state
