+++ src/components/Journal/style.css
