+++ src/components/Photos/style.css
